Walk directory for needed paths (#103)
Ref #102 

Depending on the structure of the action repository, the `node_modules`
and `package.json` needed by `local-action` may not exist in the
expected locations. This PR updates the behavior of the action to
instead walk the directory structure and attempt to locate the paths
more dynamically, allowing support for more customized action
repositories.